fixed provide json Template error
