+++ src/feature/createApi/hooks/useCreateApiForm.ts
@@ -1,8 +1,11 @@
-import {  apiFormSchema } from "@/schemas/createApiSchema";
+import { ApiFormDataType, apiFormSchema } from "@/schemas/createApiSchema";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
+import { SubmitHandler, useForm } from "react-hook-form";
 import { useCreateApiStorevalue } from "./useCreateApiStore";
 import toast from "react-hot-toast";
+import { createGhostApi } from "../mutations/createApi";
+import { ResolveValueInput } from "@/lib/faker";
+import { useRouter } from "next/navigation";
 
 // example data
 export const exampleData = {
@@ -19,9 +22,11 @@ export const exampleData = {
 };
 
 export const useCreateApiForm = () => {
+  const router = useRouter();
   const {
     register,
     handleSubmit,
+    getValues,
     watch,
     setValue,
     reset,
@@ -79,6 +84,80 @@ export const useCreateApiForm = () => {
   const apiMethod = watch("apiMethod");
   const jsonTemplate = watch("jsonTemplate");
 
+  const {
+    setIsAiGenerating,
+    aiPrompt,
+    setAiPrompt,
+
+    setTab,
+    setIsLoading,
+  } = useCreateApiStorevalue();
+
+  const generateAiJson = async () => {
+    if (!aiPrompt.trim()) {
+      toast.error("Please describe what data you need");
+      return;
+    }
+
+    setIsAiGenerating(true);
+    try {
+      const data = {
+        userMessage: aiPrompt,
+      };
+      const response = await fetch("/api/v1/ai-generated", {
+        method: "POST",
+        body: JSON.stringify(data),
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+
+      const result = await response.json();
+
+      const jsonData = JSON.parse(result.data);
+
+      const aiGeneratedData = jsonData.schema as [
+        key: string,
+        ResolveValueInput
+      ];
+
+      setValue("jsonTemplate", aiGeneratedData);
+      console.log(getValues("jsonTemplate"));
+
+      setJsonString(JSON.stringify(aiGeneratedData, null, 2));
+      toast.success("AI generated JSON template!");
+      setTab("json");
+
+      setAiPrompt("");
+    } catch (error) {
+      toast.error("Failed to generate JSON");
+    } finally {
+      setIsAiGenerating(false);
+    }
+  };
+
+  const onSubmit: SubmitHandler<ApiFormDataType> = async (data) => {
+    if (!data.jsonTemplate || Object.keys(data.jsonTemplate).length === 0) {
+      toast.error("Please provide a JSON template");
+      return;
+    }
+
+    setIsLoading(true);
+    try {
+      await createGhostApi(data);
+      toast.success("API endpoint created successfully!");
+      reset();
+      setJsonString("");
+      router.push("/my-api");
+    } catch (error) {
+      toast.error("Failed to create API endpoint");
+      console.error(error);
+    } finally {
+      reset();
+      setIsLoading(false);
+    }
+  };
+
   return {
     // react-hook-form
     register,
@@ -90,6 +169,7 @@ export const useCreateApiForm = () => {
     defaultCount,
     apiMethod,
     jsonTemplate,
+    getValues,
 
     // handlers
     handleArrayCountValue,
@@ -97,5 +177,7 @@ export const useCreateApiForm = () => {
     clearJsonTemplate,
     formatJson,
     handleJsonChange,
+    onSubmit,
+    generateAiJson,
   };
 };
